feat(frontend): extract TTS logic to useTts hook and add TTS to main page agent
